Wrote python code to interact with DynamoDB.
Added keg_interface.py, which listens to the Arduino over serial.  When drink information is sent, it gets parsed and stored in DynamoDB.  The individual drinks are recorded, and the user's lifetime drink totals get updated to ease in the website's job of displaying the data.
Still need to write the parser that reads from DynamoDB and generates the static web page every so often.  That seems easier than having it be dynamic, so I'll do the static version first.