Initial static webpage creator.
data_scraper.py pulls drinker info from DynamoDB and fills out a template, which it then uploads to S3 to be viewed as a static webpage.  Currently the scraper must be run manually, but at some point there'll be an EC2 instance (or some other machine) running it as a cron.